Prevents another users to perform updation & deletion in subtasks

+++ src/services/subtasksService.js
@@ -22,6 +22,23 @@ class SubTaskService {
     }
 
     async updateSubTask (subtaskId, subtask, userId) {
+        const currentSubtask = await this.subtaskRepository.findSubTaskById(subtaskId);
+
+        if (!currentSubtask) {
+            const err = new Error("Subtask not found");
+            err.statusCode = 404;
+            throw err;
+        }
+
+        // Only creator or assignedUser can update
+        if (
+            currentSubtask.createdBy._id.toString() !== userId.toString() &&
+            (!currentSubtask.assignedUser || currentSubtask.assignedUser._id.toString() !== userId.toString())
+        ) {
+            const err = new Error("You are not authorized to update this subtask.");
+            err.statusCode = 403;
+            throw err;
+        }
         const updated = await this.subtaskRepository.updateSubTask(subtaskId, subtask);
 
         this.io.emit("subtaskUpdated", updated);
@@ -31,6 +48,24 @@ class SubTaskService {
     }
 
     async deleteSubTasks (subtaskId, userId) {
+        const currentSubtask = await this.subtaskRepository.findSubTaskById(subtaskId);
+
+        if (!currentSubtask) {
+            const err = new Error("Subtask not found");
+            err.statusCode = 404;
+            throw err;
+        }
+
+    // Only creator or assignedUser can delete
+        if (
+            currentSubtask.createdBy._id.toString() !== userId.toString() &&
+            (!currentSubtask.assignedUser || currentSubtask.assignedUser._id.toString() !== userId.toString())
+        ) {
+            const err = new Error("You are not authorized to delete this subtask.");
+            err.statusCode = 403;
+            throw err;
+        }
+        
         const deleted = await this.subtaskRepository.deleteSubTask(subtaskId);
 
         this.io.emit("subtaskDeleted", deleted);
